refactor(ui): reorganize DropIndicator import path for better structure 🔄📁
